Refactor evaluation system to use Strategy Pattern with pluggable evaluators
