+++ evaluations/services/evaluators/__init__.py
@@ -1,0 +1,15 @@
+"""Evaluator classes for different evaluation types."""
+
+from evaluations.services.evaluators.base_evaluator import BaseEvaluator
+from evaluations.services.evaluators.first_test_evaluator import FirstTestEvaluator
+from evaluations.services.evaluators.models_evaluator import ModelsEvaluator
+from evaluations.services.evaluators.additional_tests_evaluator import AdditionalTestsEvaluator
+from evaluations.services.evaluators.additional_models_evaluator import AdditionalModelsEvaluator
+
+__all__ = [
+    "BaseEvaluator",
+    "FirstTestEvaluator",
+    "ModelsEvaluator",
+    "AdditionalTestsEvaluator",
+    "AdditionalModelsEvaluator",
+]
